backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pymongo==4.13.2
motor==3.7.1
pydantic==2.11.7
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1
aiofiles==24.1.0
Pillow==11.3.0
openai==1.97.1
cryptography==45.0.5
email-validator==2.2.0
httpx==0.28.1
